sds/src/main/java/ch/cyberduck/core/sds/SDSTouchFeature.java
package ch.cyberduck.core.sds;

/*
 * Copyright (c) 2002-2017 iterate GmbH. All rights reserved.
 * https://cyberduck.io/
 *
 * This program is free software; you can redistribute it and/or modify
 * it under the terms of the GNU General Public License as published by
 * the Free Software Foundation; either version 2 of the License, or
 * (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * GNU General Public License for more details.
 */

import ch.cyberduck.core.Path;
import ch.cyberduck.core.VersionId;
import ch.cyberduck.core.shared.DefaultTouchFeature;
import ch.cyberduck.core.shared.DefaultUploadFeature;

public class SDSTouchFeature extends DefaultTouchFeature<VersionId> {

    public SDSTouchFeature(final SDSSession session) {
        super(new DefaultUploadFeature<VersionId>(new SDSWriteFeature(session)));
    }

    @Override
    public boolean isSupported(final Path workdir) {
        return !workdir.isRoot();
    }
}
